Beginning work on cogs. Hopefully everything holds up...
