+++ nyxsplash.py
@@ -1,39 +1,76 @@
 # Splash image for console screens. No good reason for it other than it looks cool.
 
-print("                 \033[37m=")
-print("               \033[37m \033[1m\033[36m=\033[33mII\033[22m")
-print("            \033[37m,\033[1m=\033[33mII\033[37mII\033[33m=\033[22m\033[37m=                \033[34m")
-print("         \033[34m \033[1m\033[33m=II\033[37mI\033[22mI~\033[34m::++\033[1m+\033[22m            \033[34m \033[1m\033[36m=\033[37mI\033[22m \033[34m     \033[34m")
-print("       \033[34m \033[37m=\033[1m\033[33mIII\033[22m\033[37mI\033[1m\033[34m:\033[22m\033[37m==\033[35m,\033[1m\033[34m,=\033[22m+=\033[1m=\033[22m          \033[34m \033[1m\033[33mII\033[22m\033[34m       \033[37m \033[1m\033[34m=\033[22m")
-print("       \033[1m\033[33mII\033[37m7\033[36mI\033[34m++\033[22m\033[37m=\033[1m\033[33mIII=7=\033[34m===\033[22m         \033[34m \033[1m\033[33mII\033[36m=\033[22m\033[34m     \033[34m \033[1m\033[36m=I\033[22m")
-print("  \033[37m:7\033[1m\033[33mI\033[22m\033[37mII\033[1m\033[33mII\033[36mII\033[34m+=\033[33mIIIII\033[22m=7\033[1m\033[37m=\033[34m==\033[22m          \033[1m\033[36mIIII===I\033[37mI\033[36mI\033[22m\033[37m,,,")
-print("      \033[37m \033[1m\033[33mI7\033[36m7I\033[34m=7\033[37mI\033[33m7==\033[22m=\033[1m7I\033[37mI\033[34m=I=\033[22m      \033[37m \033[1m=\033[22m=,\033[1mI\033[33mI\033[37mI\033[36mII=\033[34mI\033[22m\033[37m7 \033[33m \033[37m, \033[37m=,\033[1mI\033[22m")
-print("       \033[37m \033[1m\033[33m77\033[37m=\033[36mI\033[34m=7\033[37mI\033[33mIIII\033[37m=\033[33mI\033[37m7\033[22m=~   \033[34m \033[37m+\033[33m \033[37m=\033[33m  \033[1mIIIII\033[37mI\033[33m7\033[22m\033[37m7\033[1m\033[33m=I\033[22m\033[37m~I7~=")
-print("     \033[37m \033[1m=\033[22m=\033[1m\033[33mI\033[22m7\033[1mIIII\033[37m7\033[36mI\033[37mII\033[36mII\033[22m\033[37mI\033[1m\033[33m7\033[22m\033[37mI\033[33m=    \033[33m I\033[37m7 7\033[1m\033[33mI\033[37m7\033[33mI\033[37m7I\033[35mI\033[37m7\033[36m7\033[33mI\033[22m\033[37m= \033[37m \033[1m\033[35m=\033[22m \033[37m7\033[1m\033[35m=\033[22m\033[34m")
-print("     \033[37m  \033[37m  \033[37m,\033[1m\033[33m7\033[22m\033[37mI \033[34m \033[1m\033[33m==\033[34m:~~=\033[22m         \033[1m\033[35mI\033[22m\033[34m \033[37m~~\033[1m\033[33mI\033[37m=I\033[33mII\033[37mI==\033[33mII\033[22m\033[37m  \033[37mI= \033[37m \033[1m\033[36mI\033[22m     \033[34m")
-print("        \033[34m \033[37mI   \033[33m \033[1m\033[37mI\033[22m\033[34m \033[1mI\033[36m7\033[34m7I\033[22m\033[33m \033[37m~\033[33m \033[37m==,  \033[34m \033[1m\033[37mI\033[22m, \033[37m7\033[1m\033[33mI\033[37mII\033[22m7\033[1m\033[33m=\033[22m\033[37m=,\033[1mI7\033[33m7I\033[22m\033[37m=,7       \033[34m \033[1m\033[36mIIII=\033[22m\033[37m,\033[34m")
-print("            \033[34m \033[37m=:, \033[1m\033[34m7I=I\033[36m==\033[22m\033[37m=\033[1m=\033[22m   \033[34m  \033[37m= \033[1m\033[37mIII\033[33m7IIII7\033[37mII\033[33mII\033[37m=\033[22m7I,      \033[34m \033[1m\033[36m7IIIII=\033[34m=\033[22m")
-print("          \033[37m :77\033[1m7\033[22m=I=\033[1m\033[36m=\033[37mI7\033[22mI \033[1m\033[34m7=\033[22m     \033[34m \033[37m=\033[1mI\033[22mI\033[1m\033[33mI\033[22m\033[37m=\033[1mI\033[33mI\033[37mI\033[22mI\033[1m\033[34m=\033[33mI\033[34mI\033[37mI\033[33mIII\033[22m\033[37m,\033[34m \033[37m+     \033[34m,\033[1m=\033[22m+ \033[1m\033[36mIII7==\033[34m7\033[22m\033[37m,\033[34m")
-print("              \033[34m \033[1m=\033[36m=7III\033[22m\033[34m   7\033[1m=\033[22m,  \033[37m,=I7,\033[1m\033[34mI\033[22m\033[37m=7=\033[1m\033[34m7\033[22m\033[37m=\033[1m\033[34m7\033[22m\033[37m=\033[1m\033[34mI\033[36m7\033[37mI\033[33mII7\033[37m=\033[22m,\033[34m   \033[34m,\033[1m=7\033[22m+   7\033[1m\033[36mI77\033[34m=\033[22m7\033[1m=7\033[22m,")
-print("            \033[34m ,\033[1m\033[36m=\033[34m7=I\033[22m\033[37m,\033[1m\033[36mI\033[22m\033[34m~     :\033[1m=\033[37mI=\033[22m=\033[1m\033[34m=\033[22m\033[37mII=\033[1m\033[36m=\033[22m\033[37mI\033[1m\033[36mI=\033[34m=\033[22m\033[37mI=777\033[1m=7\033[33mI\033[36m7\033[33mI\033[36m=\033[22m\033[34m~\033[1m7\033[22m:       +\033[1m\033[36m=\033[34m=7=\033[22m+\033[1m7=\033[22m7,")
-print("           \033[34m \033[1mI\033[22m7=\033[1m=I\033[22m \033[37mI\033[1m\033[36mI\033[34m=,,==\033[22m\033[37m=\033[1m\033[36m7\033[37m==\033[22m=\033[1m=\033[22m==\033[1m=\033[33mI=\033[22m\033[37m7\033[1m\033[36m=\033[22m\033[37m=\033[1m\033[33m==\033[36m7\033[22m\033[37m=7\033[1m=\033[22m7\033[1m=I=\033[36m7\033[33mI7\033[34m=\033[22m:      : \033[1m7\033[36m7\033[22m\033[34m:\033[1m7\033[22m=\033[1m=7=\033[22m 7")
-print("          \033[34m \033[1m\033[36m7==\033[34m7I\033[22m  :\033[1mI\033[22m\033[37m=\033[34m+ ,,\033[1m=7I\033[22m\033[37m77,\033[1m\033[36m=\033[37mI\033[22m7\033[1m\033[35m7\033[37m=\033[35m=\033[22m\033[37m~\033[1mI=\033[33m7\033[22m\033[37mI\033[1mI\033[36m=\033[22m\033[37m,7=\033[1m\033[34m7\033[22m\033[37mI\033[1m\033[34m=7\033[33mII\033[36m=\033[34m,\033[22m+ :+::+~\033[1m\033[36m=\033[22m\033[34m=\033[1m==\033[36m=\033[34m=\033[22m7\033[1m=\033[22m \033[1m+\033[22m")
-print("        \033[37m \033[34m~\033[1mI\033[36mI7=\033[34mI\033[22m~ :  \033[1m\033[36mI\033[22m\033[34m + \033[1m\033[36m=\033[34m,7\033[37m7II\033[33m=\033[22m7\033[37m \033[34m \033[1m\033[37mI\033[33mII\033[22m\033[37m=\033[1m\033[33mIII\033[22m\033[37m:\033[1m\033[33mI7=\033[37mI7I=\033[34m+7\033[37m7\033[33mII\033[36m==\033[22m\033[34m,==, : \033[1m\033[36m7\033[34m=\033[36m==\033[34mI=\033[22m :  \033[1m7\033[22m")
-print("       \033[34m \033[1m=\033[22m~\033[1m7\033[36mIIII\033[22m\033[34m  : =\033[37mI\033[1m\033[34m7\033[22m \033[37mII\033[1mI\033[36mI\033[33mI=I\033[34m=\033[22m\033[37m,\033[34m:  \033[37m7\033[1m\033[36m7\033[22m\033[37m=\033[1m\033[33mII\033[34m \033[22m :\033[37m7\033[1m\033[33mIII==\033[37mI\033[36m7I7II=\033[34m7\033[36m==\033[34m=\033[22m,+~\033[1mI\033[36m77I\033[22m\033[34m:\033[1m=\033[22m     :,")
-print("      \033[34m \033[1m=\033[22m  \033[1m\033[36m=\033[22m\033[34m=\033[1m\033[36mI7\033[34m=\033[22m :~~ \033[37m \033[1m\033[36m7\033[22m\033[37m7\033[1m7\033[33m==I\033[34m====\033[22m\033[37m=\033[1m\033[34m=\033[22m  \033[37m7I7\033[1m\033[33mI\033[37mI\033[22m\033[34m   \033[1m=\033[37mI\033[33mI=\033[22m\033[34m~\033[1m\033[33m7===I\033[37m7\033[36m77=\033[22m\033[34m 7\033[1m\033[36mI\033[34m=\033[22m :\033[1m\033[36m=7II\033[22m\033[34m \033[1m=\033[22m  :")
-print("      \033[34m:::\033[1m7\033[36mI\033[34m7\033[36m7I\033[22m\033[34m  +    \033[1m\033[36mI7\033[37mI\033[33mII\033[34m+7=\033[22m:\033[1m\033[33mI\033[36m=\033[34m=\033[22m   \033[37m=7I\033[1m\033[34m=\033[22m   \033[1m7\033[37mI\033[33mII\033[34m=\033[22m:\033[37m:7\033[1m\033[33mI=77\033[36mI\033[34m7\033[36m=\033[22m\033[34m  :\033[1m7\033[22m=\033[1m\033[36m7II\033[34mI\033[22m \033[1mI\033[22m  :~  :")
-print("     \033[34m  ::\033[1m=\033[22m \033[1mI\033[36m7I\033[22m\033[34m   \033[37m  \033[34m \033[1m\033[36mIIII=\033[22m\033[37m7,=\033[34m::\033[1m\033[33mII\033[36m=\033[22m\033[34m:  \033[37m~\033[1m\033[34m=\033[22m\033[37m==\033[34m ::\033[1m\033[36m=\033[33mIII\033[34m=\033[22m::\033[37m=7I\033[1m\033[35m=\033[36mII\033[22m\033[34m7\033[1m7\033[36m=\033[22m\033[34m  \033[34m \033[1m\033[36m7II7\033[22m\033[34m~     ~     \033[34m")
-print("      \033[34m   \033[1m=\033[22m \033[1mI\033[36m7I\033[22m    \033[34m \033[1m\033[36mI\033[22m\033[34m,\033[1m\033[36m77II\033[34m=\033[22m\033[33m=\033[37m=,\033[34m=\033[1m\033[36m7\033[33mIII\033[37m=\033[22m=\033[1m\033[34m=7\033[22m\033[35m7\033[37m,=\033[1m\033[33m=IIII\033[37mI\033[22m\033[34m=+\033[37m+=\033[1m=\033[34m=\033[36mI\033[37mI\033[36mI\033[34m=I\033[22m7\033[1m\033[36mI\033[22m\033[34m  \033[1m\033[36mII\033[22m\033[34m=\033[1mI\033[22m   ::")
-print("      \033[34m : = :\033[1m\033[36m=I\033[22m    \033[1m\033[36mI\033[22m\033[34m \033[1m\033[36mIII\033[34mI\033[22m\033[37m,\033[1m\033[36mI\033[22m\033[37m=+\033[1mI\033[22m\033[34m:+\033[1m\033[36m7\033[37mI\033[33mIIII\033[37mI\033[22m=7\033[1m\033[33mIIIII\033[36mI\033[34mI\033[22m=\033[37m=\033[1m\033[34m7=\033[36m==\033[34m=\033[36mII==\033[22m\033[34m \033[1m7\033[22m=\033[1m\033[36mI7\033[22m\033[34m~\033[1m7\033[22m :: ::   :")
-print("      \033[37m \033[34m=  : \033[1m7\033[36mI\033[22m    \033[1m\033[36mI\033[22m \033[1m\033[36mIII\033[22m\033[34m \033[1m\033[33mI\033[36m==\033[34m==\033[33mI\033[22m\033[37m=\033[1m\033[34m==\033[36m7I\033[37mI\033[33mII\033[22m\033[37m=\033[1mI\033[33mI\033[37mII\033[36m7\033[34mI\033[22m=\033[1m=\033[22m\033[37m7+\033[1m\033[33m7\033[36m=77\033[37mI\033[36mI=II\033[22m \033[1m\033[34mI7\033[36m=\033[22m\033[34m7  ::  :   \033[34m")
-print("       \033[34m~  :  \033[1m\033[36m7\033[22m\033[34m    \033[37m: \033[1m\033[34mI\033[36m=\033[34m7\033[22m\033[37m~\033[34mI\033[1m\033[36m77\033[37mI\033[34m=\033[36m7\033[33m=\033[22m\033[37m=7=\033[1m\033[34m=\033[22m\033[37m=\033[1m\033[34m=====\033[36m=\033[22m\033[37m=\033[1m\033[34m=\033[22m\033[37mI\033[1m\033[33mI\033[34m=\033[33mI\033[36m=7\033[37mI\033[36m77\033[34mI7\033[36m77\033[34mI\033[22m \033[1m=7\033[22m    :")
-print("       \033[34m   ::        \033[34m \033[1m\033[36m7\033[22m\033[34m \033[1m\033[36m=\033[22m\033[34m \033[1m7\033[37mI\033[36m=\033[34m=\033[36m===I\033[37mI\033[22m=\033[1m\033[33m===7\033[22m\033[37m=\033[1m\033[33mI\033[36m=\033[33mI\033[22m\033[37m=\033[1m\033[33mI\033[36m===77=I\033[34m=\033[22m \033[1m\033[36mI\033[34m77\033[22m=:  ::")
-print("       \033[34m    :         \033[34m7 \033[1m\033[36mI\033[22m\033[34m :\033[1m\033[36mI=\033[34m7\033[33mI\033[37mI\033[36mI==I7==\033[22m\033[37m=\033[1m\033[36m==7\033[34m=\033[36m77I\033[34m=\033[33mII\033[34m=\033[36mI=\033[22m\033[34m \033[1m\033[36mI\033[34m77\033[22m~  \033[34m ::     \033[34m")
-print("          \033[34m  :         \033[34m:\033[1m\033[36mII\033[22m\033[34m  \033[1mI\033[36m=\033[34m=\033[22m,7\033[1m\033[36m7\033[33mII\033[34m=\033[36m7I\033[33mI\033[36m=II\033[34m=\033[33mII\033[36m=II\033[34m=\033[36m=\033[34m7\033[22m,\033[1m\033[36mI\033[34m7\033[22m~\033[1mI\033[22m  \033[34m ::     \033[34m")
-print("           \033[34m           \033[37m=\033[1mI\033[36mII\033[22m\033[34m  \033[34m \033[1m77=\033[22m \033[1m7\033[36m==\033[34m=\033[36mI\033[33mII\033[36mI\033[34mI==\033[36mI\033[34m=\033[36mI\033[34m=\033[36m7\033[34m7\033[22m,\033[1m\033[36m7\033[22m\033[34m:  \033[34m    :")
-print("            \033[34m          \033[34m \033[1m7\033[36m7\033[34m~\033[22m    \033[34m 7\033[1m=\033[22m  \033[1mI\033[36mI\033[34m=\033[35mI\033[36mI\033[34mI\033[22m7=\033[1m==7=7\033[22m  \033[1m7\033[22m     \033[34m    \033[37m")
-print("              \033[34m          \033[37m7\033[34m       \033[34m~= \033[37m \033[34m7\033[1m\033[36m7\033[34m=\033[22m7 ~=\033[1m==7\033[22m   \033[34m:     \033[34m")
-print("                       \033[37m \033[1m\033[34m=\033[36m7\033[22m\033[37m,\033[34m     \033[34m     \033[34m:\033[1m\033[36m7\033[22m\033[34m  \033[34m+\033[1m7\033[22m:     \033[34m     \033[34m")
-print("                       \033[37mI\033[1m\033[33mI\033[22m\033[37m \033[33m \033[37m:         \033[34m \033[1m7\033[22m =:+")
-print("                       \033[37m 7~           \033[34m  \033[34m")
-print("\033[35m                                 __     __)\n                                (, /|  /\n                                  / | /      __/\n                               ) /  |/   (_/_ /(__\n                              (_/   '   .-/  /\n                                       (_/\033[39m")
+def splash():
+    print("                 \033[37m=")
+    print("               \033[37m \033[1m\033[36m=\033[33mII\033[22m")
+    print("            \033[37m,\033[1m=\033[33mII\033[37mII\033[33m=\033[22m\033[37m=                \033[34m")
+    print(
+        "         \033[34m \033[1m\033[33m=II\033[37mI\033[22mI~\033[34m::++\033[1m+\033[22m            \033[34m \033[1m\033[36m=\033[37mI\033[22m \033[34m     \033[34m")
+    print(
+        "       \033[34m \033[37m=\033[1m\033[33mIII\033[22m\033[37mI\033[1m\033[34m:\033[22m\033[37m==\033[35m,\033[1m\033[34m,=\033[22m+=\033[1m=\033[22m          \033[34m \033[1m\033[33mII\033[22m\033[34m       \033[37m \033[1m\033[34m=\033[22m")
+    print(
+        "       \033[1m\033[33mII\033[37m7\033[36mI\033[34m++\033[22m\033[37m=\033[1m\033[33mIII=7=\033[34m===\033[22m         \033[34m \033[1m\033[33mII\033[36m=\033[22m\033[34m     \033[34m \033[1m\033[36m=I\033[22m")
+    print(
+        "  \033[37m:7\033[1m\033[33mI\033[22m\033[37mII\033[1m\033[33mII\033[36mII\033[34m+=\033[33mIIIII\033[22m=7\033[1m\033[37m=\033[34m==\033[22m          \033[1m\033[36mIIII===I\033[37mI\033[36mI\033[22m\033[37m,,,")
+    print(
+        "      \033[37m \033[1m\033[33mI7\033[36m7I\033[34m=7\033[37mI\033[33m7==\033[22m=\033[1m7I\033[37mI\033[34m=I=\033[22m      \033[37m \033[1m=\033[22m=,\033[1mI\033[33mI\033[37mI\033[36mII=\033[34mI\033[22m\033[37m7 \033[33m \033[37m, \033[37m=,\033[1mI\033[22m")
+    print(
+        "       \033[37m \033[1m\033[33m77\033[37m=\033[36mI\033[34m=7\033[37mI\033[33mIIII\033[37m=\033[33mI\033[37m7\033[22m=~   \033[34m \033[37m+\033[33m \033[37m=\033[33m  \033[1mIIIII\033[37mI\033[33m7\033[22m\033[37m7\033[1m\033[33m=I\033[22m\033[37m~I7~=")
+    print(
+        "     \033[37m \033[1m=\033[22m=\033[1m\033[33mI\033[22m7\033[1mIIII\033[37m7\033[36mI\033[37mII\033[36mII\033[22m\033[37mI\033[1m\033[33m7\033[22m\033[37mI\033[33m=    \033[33m I\033[37m7 7\033[1m\033[33mI\033[37m7\033[33mI\033[37m7I\033[35mI\033[37m7\033[36m7\033[33mI\033[22m\033[37m= \033[37m \033[1m\033[35m=\033[22m \033[37m7\033[1m\033[35m=\033[22m\033[34m")
+    print(
+        "     \033[37m  \033[37m  \033[37m,\033[1m\033[33m7\033[22m\033[37mI \033[34m \033[1m\033[33m==\033[34m:~~=\033[22m         \033[1m\033[35mI\033[22m\033[34m \033[37m~~\033[1m\033[33mI\033[37m=I\033[33mII\033[37mI==\033[33mII\033[22m\033[37m  \033[37mI= \033[37m \033[1m\033[36mI\033[22m     \033[34m")
+    print(
+        "        \033[34m \033[37mI   \033[33m \033[1m\033[37mI\033[22m\033[34m \033[1mI\033[36m7\033[34m7I\033[22m\033[33m \033[37m~\033[33m \033[37m==,  \033[34m \033[1m\033[37mI\033[22m, \033[37m7\033[1m\033[33mI\033[37mII\033[22m7\033[1m\033[33m=\033[22m\033[37m=,\033[1mI7\033[33m7I\033[22m\033[37m=,7       \033[34m \033[1m\033[36mIIII=\033[22m\033[37m,\033[34m")
+    print(
+        "            \033[34m \033[37m=:, \033[1m\033[34m7I=I\033[36m==\033[22m\033[37m=\033[1m=\033[22m   \033[34m  \033[37m= \033[1m\033[37mIII\033[33m7IIII7\033[37mII\033[33mII\033[37m=\033[22m7I,      \033[34m \033[1m\033[36m7IIIII=\033[34m=\033[22m")
+    print(
+        "          \033[37m :77\033[1m7\033[22m=I=\033[1m\033[36m=\033[37mI7\033[22mI \033[1m\033[34m7=\033[22m     \033[34m \033[37m=\033[1mI\033[22mI\033[1m\033[33mI\033[22m\033[37m=\033[1mI\033[33mI\033[37mI\033[22mI\033[1m\033[34m=\033[33mI\033[34mI\033[37mI\033[33mIII\033[22m\033[37m,\033[34m \033[37m+     \033[34m,\033[1m=\033[22m+ \033[1m\033[36mIII7==\033[34m7\033[22m\033[37m,\033[34m")
+    print(
+        "              \033[34m \033[1m=\033[36m=7III\033[22m\033[34m   7\033[1m=\033[22m,  \033[37m,=I7,\033[1m\033[34mI\033[22m\033[37m=7=\033[1m\033[34m7\033[22m\033[37m=\033[1m\033[34m7\033[22m\033[37m=\033[1m\033[34mI\033[36m7\033[37mI\033[33mII7\033[37m=\033[22m,\033[34m   \033[34m,\033[1m=7\033[22m+   7\033[1m\033[36mI77\033[34m=\033[22m7\033[1m=7\033[22m,")
+    print(
+        "            \033[34m ,\033[1m\033[36m=\033[34m7=I\033[22m\033[37m,\033[1m\033[36mI\033[22m\033[34m~     :\033[1m=\033[37mI=\033[22m=\033[1m\033[34m=\033[22m\033[37mII=\033[1m\033[36m=\033[22m\033[37mI\033[1m\033[36mI=\033[34m=\033[22m\033[37mI=777\033[1m=7\033[33mI\033[36m7\033[33mI\033[36m=\033[22m\033[34m~\033[1m7\033[22m:       +\033[1m\033[36m=\033[34m=7=\033[22m+\033[1m7=\033[22m7,")
+    print(
+        "           \033[34m \033[1mI\033[22m7=\033[1m=I\033[22m \033[37mI\033[1m\033[36mI\033[34m=,,==\033[22m\033[37m=\033[1m\033[36m7\033[37m==\033[22m=\033[1m=\033[22m==\033[1m=\033[33mI=\033[22m\033[37m7\033[1m\033[36m=\033[22m\033[37m=\033[1m\033[33m==\033[36m7\033[22m\033[37m=7\033[1m=\033[22m7\033[1m=I=\033[36m7\033[33mI7\033[34m=\033[22m:      : \033[1m7\033[36m7\033[22m\033[34m:\033[1m7\033[22m=\033[1m=7=\033[22m 7")
+    print(
+        "          \033[34m \033[1m\033[36m7==\033[34m7I\033[22m  :\033[1mI\033[22m\033[37m=\033[34m+ ,,\033[1m=7I\033[22m\033[37m77,\033[1m\033[36m=\033[37mI\033[22m7\033[1m\033[35m7\033[37m=\033[35m=\033[22m\033[37m~\033[1mI=\033[33m7\033[22m\033[37mI\033[1mI\033[36m=\033[22m\033[37m,7=\033[1m\033[34m7\033[22m\033[37mI\033[1m\033[34m=7\033[33mII\033[36m=\033[34m,\033[22m+ :+::+~\033[1m\033[36m=\033[22m\033[34m=\033[1m==\033[36m=\033[34m=\033[22m7\033[1m=\033[22m \033[1m+\033[22m")
+    print(
+        "        \033[37m \033[34m~\033[1mI\033[36mI7=\033[34mI\033[22m~ :  \033[1m\033[36mI\033[22m\033[34m + \033[1m\033[36m=\033[34m,7\033[37m7II\033[33m=\033[22m7\033[37m \033[34m \033[1m\033[37mI\033[33mII\033[22m\033[37m=\033[1m\033[33mIII\033[22m\033[37m:\033[1m\033[33mI7=\033[37mI7I=\033[34m+7\033[37m7\033[33mII\033[36m==\033[22m\033[34m,==, : \033[1m\033[36m7\033[34m=\033[36m==\033[34mI=\033[22m :  \033[1m7\033[22m")
+    print(
+        "       \033[34m \033[1m=\033[22m~\033[1m7\033[36mIIII\033[22m\033[34m  : =\033[37mI\033[1m\033[34m7\033[22m \033[37mII\033[1mI\033[36mI\033[33mI=I\033[34m=\033[22m\033[37m,\033[34m:  \033[37m7\033[1m\033[36m7\033[22m\033[37m=\033[1m\033[33mII\033[34m \033[22m :\033[37m7\033[1m\033[33mIII==\033[37mI\033[36m7I7II=\033[34m7\033[36m==\033[34m=\033[22m,+~\033[1mI\033[36m77I\033[22m\033[34m:\033[1m=\033[22m     :,")
+    print(
+        "      \033[34m \033[1m=\033[22m  \033[1m\033[36m=\033[22m\033[34m=\033[1m\033[36mI7\033[34m=\033[22m :~~ \033[37m \033[1m\033[36m7\033[22m\033[37m7\033[1m7\033[33m==I\033[34m====\033[22m\033[37m=\033[1m\033[34m=\033[22m  \033[37m7I7\033[1m\033[33mI\033[37mI\033[22m\033[34m   \033[1m=\033[37mI\033[33mI=\033[22m\033[34m~\033[1m\033[33m7===I\033[37m7\033[36m77=\033[22m\033[34m 7\033[1m\033[36mI\033[34m=\033[22m :\033[1m\033[36m=7II\033[22m\033[34m \033[1m=\033[22m  :")
+    print(
+        "      \033[34m:::\033[1m7\033[36mI\033[34m7\033[36m7I\033[22m\033[34m  +    \033[1m\033[36mI7\033[37mI\033[33mII\033[34m+7=\033[22m:\033[1m\033[33mI\033[36m=\033[34m=\033[22m   \033[37m=7I\033[1m\033[34m=\033[22m   \033[1m7\033[37mI\033[33mII\033[34m=\033[22m:\033[37m:7\033[1m\033[33mI=77\033[36mI\033[34m7\033[36m=\033[22m\033[34m  :\033[1m7\033[22m=\033[1m\033[36m7II\033[34mI\033[22m \033[1mI\033[22m  :~  :")
+    print(
+        "     \033[34m  ::\033[1m=\033[22m \033[1mI\033[36m7I\033[22m\033[34m   \033[37m  \033[34m \033[1m\033[36mIIII=\033[22m\033[37m7,=\033[34m::\033[1m\033[33mII\033[36m=\033[22m\033[34m:  \033[37m~\033[1m\033[34m=\033[22m\033[37m==\033[34m ::\033[1m\033[36m=\033[33mIII\033[34m=\033[22m::\033[37m=7I\033[1m\033[35m=\033[36mII\033[22m\033[34m7\033[1m7\033[36m=\033[22m\033[34m  \033[34m \033[1m\033[36m7II7\033[22m\033[34m~     ~     \033[34m")
+    print(
+        "      \033[34m   \033[1m=\033[22m \033[1mI\033[36m7I\033[22m    \033[34m \033[1m\033[36mI\033[22m\033[34m,\033[1m\033[36m77II\033[34m=\033[22m\033[33m=\033[37m=,\033[34m=\033[1m\033[36m7\033[33mIII\033[37m=\033[22m=\033[1m\033[34m=7\033[22m\033[35m7\033[37m,=\033[1m\033[33m=IIII\033[37mI\033[22m\033[34m=+\033[37m+=\033[1m=\033[34m=\033[36mI\033[37mI\033[36mI\033[34m=I\033[22m7\033[1m\033[36mI\033[22m\033[34m  \033[1m\033[36mII\033[22m\033[34m=\033[1mI\033[22m   ::")
+    print(
+        "      \033[34m : = :\033[1m\033[36m=I\033[22m    \033[1m\033[36mI\033[22m\033[34m \033[1m\033[36mIII\033[34mI\033[22m\033[37m,\033[1m\033[36mI\033[22m\033[37m=+\033[1mI\033[22m\033[34m:+\033[1m\033[36m7\033[37mI\033[33mIIII\033[37mI\033[22m=7\033[1m\033[33mIIIII\033[36mI\033[34mI\033[22m=\033[37m=\033[1m\033[34m7=\033[36m==\033[34m=\033[36mII==\033[22m\033[34m \033[1m7\033[22m=\033[1m\033[36mI7\033[22m\033[34m~\033[1m7\033[22m :: ::   :")
+    print(
+        "      \033[37m \033[34m=  : \033[1m7\033[36mI\033[22m    \033[1m\033[36mI\033[22m \033[1m\033[36mIII\033[22m\033[34m \033[1m\033[33mI\033[36m==\033[34m==\033[33mI\033[22m\033[37m=\033[1m\033[34m==\033[36m7I\033[37mI\033[33mII\033[22m\033[37m=\033[1mI\033[33mI\033[37mII\033[36m7\033[34mI\033[22m=\033[1m=\033[22m\033[37m7+\033[1m\033[33m7\033[36m=77\033[37mI\033[36mI=II\033[22m \033[1m\033[34mI7\033[36m=\033[22m\033[34m7  ::  :   \033[34m")
+    print(
+        "       \033[34m~  :  \033[1m\033[36m7\033[22m\033[34m    \033[37m: \033[1m\033[34mI\033[36m=\033[34m7\033[22m\033[37m~\033[34mI\033[1m\033[36m77\033[37mI\033[34m=\033[36m7\033[33m=\033[22m\033[37m=7=\033[1m\033[34m=\033[22m\033[37m=\033[1m\033[34m=====\033[36m=\033[22m\033[37m=\033[1m\033[34m=\033[22m\033[37mI\033[1m\033[33mI\033[34m=\033[33mI\033[36m=7\033[37mI\033[36m77\033[34mI7\033[36m77\033[34mI\033[22m \033[1m=7\033[22m    :")
+    print(
+        "       \033[34m   ::        \033[34m \033[1m\033[36m7\033[22m\033[34m \033[1m\033[36m=\033[22m\033[34m \033[1m7\033[37mI\033[36m=\033[34m=\033[36m===I\033[37mI\033[22m=\033[1m\033[33m===7\033[22m\033[37m=\033[1m\033[33mI\033[36m=\033[33mI\033[22m\033[37m=\033[1m\033[33mI\033[36m===77=I\033[34m=\033[22m \033[1m\033[36mI\033[34m77\033[22m=:  ::")
+    print(
+        "       \033[34m    :         \033[34m7 \033[1m\033[36mI\033[22m\033[34m :\033[1m\033[36mI=\033[34m7\033[33mI\033[37mI\033[36mI==I7==\033[22m\033[37m=\033[1m\033[36m==7\033[34m=\033[36m77I\033[34m=\033[33mII\033[34m=\033[36mI=\033[22m\033[34m \033[1m\033[36mI\033[34m77\033[22m~  \033[34m ::     \033[34m")
+    print(
+        "          \033[34m  :         \033[34m:\033[1m\033[36mII\033[22m\033[34m  \033[1mI\033[36m=\033[34m=\033[22m,7\033[1m\033[36m7\033[33mII\033[34m=\033[36m7I\033[33mI\033[36m=II\033[34m=\033[33mII\033[36m=II\033[34m=\033[36m=\033[34m7\033[22m,\033[1m\033[36mI\033[34m7\033[22m~\033[1mI\033[22m  \033[34m ::     \033[34m")
+    print(
+        "           \033[34m           \033[37m=\033[1mI\033[36mII\033[22m\033[34m  \033[34m \033[1m77=\033[22m \033[1m7\033[36m==\033[34m=\033[36mI\033[33mII\033[36mI\033[34mI==\033[36mI\033[34m=\033[36mI\033[34m=\033[36m7\033[34m7\033[22m,\033[1m\033[36m7\033[22m\033[34m:  \033[34m    :")
+    print(
+        "            \033[34m          \033[34m \033[1m7\033[36m7\033[34m~\033[22m    \033[34m 7\033[1m=\033[22m  \033[1mI\033[36mI\033[34m=\033[35mI\033[36mI\033[34mI\033[22m7=\033[1m==7=7\033[22m  \033[1m7\033[22m     \033[34m    \033[37m")
+    print(
+        "              \033[34m          \033[37m7\033[34m       \033[34m~= \033[37m \033[34m7\033[1m\033[36m7\033[34m=\033[22m7 ~=\033[1m==7\033[22m   \033[34m:     \033[34m")
+    print(
+        "                       \033[37m \033[1m\033[34m=\033[36m7\033[22m\033[37m,\033[34m     \033[34m     \033[34m:\033[1m\033[36m7\033[22m\033[34m  \033[34m+\033[1m7\033[22m:     \033[34m     \033[34m")
+    print(
+        "                       \033[37mI\033[1m\033[33mI\033[22m\033[37m \033[33m \033[37m:         \033[34m \033[1m7\033[22m =:+")
+    print("                       \033[37m 7~           \033[34m  \033[34m")
+    print(
+        "\033[35m                                 __     __)\n                                (, /|  /\n                                  / | /      __/\n                               ) /  |/   (_/_ /(__\n                              (_/   '   .-/  /\n                                       (_/\033[39m")
+
+
+splash()
